moved format to redbox-core
