--- changelog.d/+c36202ce.misc.md
@@ -1,2 +1,0 @@
-
-Journey tests - ensure correct route is used, as selected by keyword
